python: Drop some unused code.
