+++ python/dazl/pretty/_render_base.py
@@ -30,7 +30,6 @@ from ..damlast.daml_lf_1 import (
 from ..damlast.protocols import SymbolLookup
 from ..damlast.util import package_local_name, unpack_arrow_type
 from ..damlast.visitor import ExprVisitor, ModuleVisitor, PackageVisitor, TypeVisitor
-from ..model.definition import DamlDataType, DamlTemplate
 from ..prim import to_date, to_datetime
 from .options import PrettyOptions
 from .util import is_hidden_module_name, maybe_parentheses
@@ -206,7 +205,7 @@ class PrettyPrintBase(PackageVisitor[str], ModuleVisitor[str], ExprVisitor[str],
         return self.visit_def_template(None, def_data_type)
 
     def visit_def_template(
-        self, template: "DefTemplate", def_data_type: "Optional[DefDataType]" = None
+        self, template: "Optional[DefTemplate]", def_data_type: "Optional[DefDataType]" = None
     ) -> str:
         """
         Attempt to render the :class:`DefTemplate`. If this method successfully returns, then the
@@ -215,21 +214,6 @@ class PrettyPrintBase(PackageVisitor[str], ModuleVisitor[str], ExprVisitor[str],
         :param template: The :class:`DefTemplate` to attempt to render.
         :param def_data_type: The corresponding :class:`DefDataType` for that template (if known).
         :return: A rendered version of the template.
-        """
-        raise NotImplementedError
-
-    def visit_daml_data_type(self, data_type: DamlDataType) -> str:
-        """
-        Render the specified data type.
-
-        :param data_type:
-        :return:
-        """
-        raise NotImplementedError
-
-    def visit_daml_template(self, template: DamlTemplate) -> str:
-        """
-        Render a template instance.
         """
         raise NotImplementedError
